# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A Customer, Vehicle and Tire API in Go using Gin framework.",
        "title": "Tire Tread Application API",
        "contact": {
            "name": "Tim Seitzinger",
            "email": "tseitzinger@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/tires": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Responds with the list of all Tires as JSON.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tires"
                ],
                "summary": "Get Tires array",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Tire"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Takes a tire JSON and store in DB. Return saved JSON.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tires"
                ],
                "summary": "Store a new tire",
                "parameters": [
                    {
                        "description": "Tire JSON",
                        "name": "tire",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Tire"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Tire"
                        }
                    }
                }
            }
        },
        "/tires/{t_id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns the tire whose ID value matches the TID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tires"
                ],
                "summary": "Get single tire by v_id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "search tire by t_id",
                        "name": "t_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Tire"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete the tire whose ID value matches the ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tires"
                ],
                "summary": "Delete single tire by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "delete tire by t_id",
                        "name": "t_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Tire"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Updates the tire whose ID value matches the ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tires"
                ],
                "summary": "Update single tire by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "update tire by id",
                        "name": "t_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Tire JSON",
                        "name": "tire",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Tire"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Tire"
                        }
                    }
                }
            }
        },
        "/vehicles": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Responds with the list of all vehicles as JSON.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vehicles"
                ],
                "summary": "Get Vehicles array",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Vehicle"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Takes a vehicle JSON and store in DB. Return saved JSON.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vehicles"
                ],
                "summary": "Store a new vehicle",
                "parameters": [
                    {
                        "description": "Vehicle JSON",
                        "name": "vehicle",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Vehicle"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Vehicle"
                        }
                    }
                }
            }
        },
        "/vehicles/{v_id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns the vehicle whose ID value matches the ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vehicles"
                ],
                "summary": "Get single vehicle by v_id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "search vehicle by v_id",
                        "name": "v_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Vehicle"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete the vehicle whose ID value matches the ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vehicles"
                ],
                "summary": "Delete single vehicle by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "delete vehicle by v_id",
                        "name": "v_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Vehicle"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Updates the vehicle whose VID value matches the VID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vehicles"
                ],
                "summary": "Update single vehicle by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "update vehicle by vid",
                        "name": "v_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Vehicle JSON",
                        "name": "vehicle",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Vehicle"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Vehicle"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.Tire": {
            "type": "object",
            "properties": {
                "currentpressure": {
                    "type": "integer"
                },
                "make": {
                    "type": "string"
                },
                "maxpressure": {
                    "type": "integer"
                },
                "tid": {
                    "type": "integer"
                },
                "tireid": {
                    "type": "integer"
                },
                "tirelocation": {
                    "type": "string"
                },
                "tiremodel": {
                    "type": "string"
                },
                "tiresize": {
                    "type": "string"
                },
                "treadwear": {
                    "type": "string"
                }
            }
        },
        "models.Vehicle": {
            "type": "object",
            "properties": {
                "carmodel": {
                    "type": "string"
                },
                "make": {
                    "type": "string"
                },
                "tires": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Tire"
                    }
                },
                "vehid": {
                    "type": "integer"
                },
                "vid": {
                    "type": "integer"
                },
                "vin": {
                    "type": "string"
                },
                "year": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}